ft-793/time-service changed file structure
