+++ spec/main_spec.rb
@@ -1,5 +1,6 @@
 require 'spec_helper'
 require_relative '../time_service.rb'
+# require_relative '../main.rb'
 
 # Create TimeService for testing purposes comprised of the interface we are testing.
 class TimeService < Hoodoo::Services::Service
@@ -12,7 +13,6 @@ describe 'Time Service' do
   # Create rack builder object for the test.
   context "get /1/Time/now" do
     let(:app) {
-
       Rack::Builder.new do
         use( Hoodoo::Services::Middleware )
         run( TimeService.new )
@@ -24,7 +24,6 @@ describe 'Time Service' do
       get '/v1/time/now',
         nil,
         { 'CONTENT_TYPE' => 'application/json; charset=utf-8' }
-
       expect( last_response.status ).to eq 200
     end
 
@@ -32,12 +31,8 @@ describe 'Time Service' do
       get '/v1/time/now',
         nil,
         { 'CONTENT_TYPE' => 'application/json; charset=utf-8' }
-
-
       # get the json last_response body and create a hash from it.
       result = JSON.parse( last_response.body )
-      #   expect(result).to be_within(1).of ( Time.now.utc.iso8601 )
-
       # gets the time from 'result' hash and converts it to a datetime object
       # This will allow comparison with The Time.now object
       time = DateTime.parse( result['time'] )
